# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quark daemon API",
    "description": "Local control plane API served by quarkd on localhost.",
    "license": {
      "name": "MIT"
    },
    "version": "0.0.1"
  },
  "paths": {
    "/v1/coordinators/{id}/messages": {
      "get": {
        "tags": [
          "coordinators"
        ],
        "summary": "A Project coordinator's conversation so far. The coordinator id is the\nProject id. New entries arrive as `coordinator.message` events.",
        "operationId": "coordinator_messages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Coordinator id (the Project id)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Only entries with `id` greater than this.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "At most this many entries (default and maximum 1000).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranscriptItem"
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "coordinators"
        ],
        "summary": "Send a message to a Project's coordinator.",
        "description": "The coordinator id is the Project id. The text is typed into the\ncoordinator's live session; its reply arrives on the event stream as\n`coordinator.message` events, which also carry this message once the\nsession records it.",
        "operationId": "send_coordinator_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Coordinator id (the Project id)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoordinatorMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoordinatorMessageAccepted"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The Project has no workspace yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "Typing into the session failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "No live coordinator session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/decisions": {
      "get": {
        "tags": [
          "decisions"
        ],
        "summary": "The decisions inbox across all Projects.",
        "operationId": "list_decisions",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "description": "Only decisions in this state.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DecisionState"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Decision"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects.",
        "operationId": "list_projects",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create a Project.",
        "description": "Phase 0 records the Project in the daemon only. Seeding its engine\nworkspace arrives with the adapter write path.",
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get one Project.",
        "operationId": "get_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Configure a Project. Absent fields are left unchanged.",
        "operationId": "update_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{id}/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "The task board for one Project.",
        "operationId": "list_tasks",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Task detail.",
        "operationId": "get_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{id}/messages": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Steer a task's worker.",
        "description": "The message is durably recorded for the worker before this returns; the\nworker reads it on its next turn.",
        "operationId": "send_task_message",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendTaskMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Message recorded for the worker"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The engine refused or failed the call",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{id}/transcript": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "A worker's transcript so far, read from its harness session log. New\nentries arrive as `worker.transcript` events.",
        "operationId": "task_transcript",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Only entries with `id` greater than this.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "At most this many entries (default and maximum 1000).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TranscriptItem"
                  }
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{id}:cancel": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Cancel a task's worker.",
        "description": "Stops the worker and keeps its worktree and every uncommitted change;\nrelaunch picks the task up again.",
        "operationId": "cancel_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The worker is stopped"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The engine refused or failed the call",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{id}:relaunch": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Relaunch a task's worker in the same worktree.",
        "operationId": "relaunch_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelaunchTask"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "A new worker is running"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "The engine refused or failed the call",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CoordinatorMessage": {
        "type": "object",
        "description": "A message for a coordinator, typed into its session as if at the keyboard.",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
      "CoordinatorMessageAccepted": {
        "type": "object",
        "description": "The coordinator's session took the message. Its reply arrives as\n`coordinator.message` events, read from the session log.",
        "required": [
          "coordinator_id",
          "confirmed",
          "accepted_at"
        ],
        "properties": {
          "accepted_at": {
            "type": "string"
          },
          "confirmed": {
            "type": "boolean",
            "description": "`false` when the text was typed and submitted but the session did not\nconfirm the submit; check the transcript before sending again."
          },
          "coordinator_id": {
            "type": "string"
          }
        }
      },
...
      "Decision": {
        "type": "object",
        "description": "A question held for a person.",
        "required": [
          "id",
          "project_id",
          "question",
          "state",
          "opened_at"
        ],
        "properties": {
          "answer": {
            "type": [
              "string",
              "null"
            ]
          },
          "answered_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "answered_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "opened_at": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/DecisionState"
          },
          "task_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "DecisionState": {
        "type": "string",
        "enum": [
          "open",
          "answered"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "TranscriptItem": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TranscriptEntry"
          },
          {
            "type": "object",
            "required": [
              "id"
            ],
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64",
                "description": "The `seq` of the event that carried this entry; pass it as `after` to\npage, and use it to merge history with live events."
              }
            }
          }
        ],
        "description": "A transcript entry as listed by the history endpoints."
      },
...
    }
  },
...
}